requirements.txt
-i https://pypi.org/simple
aiofiles==0.4.0
aniso8601==7.0.0
apispec==3.1.0
apistar==0.7.2
boto3==1.10.26
boto==2.49.0
botocore==1.13.26
certifi==2019.9.11
chardet==3.0.4
click==7.0
docopt==0.6.2
docutils==0.15.2
gensim==3.8.1
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
h11==0.8.1
httptools==0.0.13 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'pypy'
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
markupsafe==1.1.1
marshmallow==3.2.2
mecab-python3==0.996.2
mypy-extensions==0.4.3
mypy==0.740
numpy==1.17.4
pandas==0.25.3
promise==2.2.1
python-dateutil==2.8.0 ; python_version >= '2.7'
python-multipart==0.0.5
pytz==2019.3
pyyaml==5.1.2
requests-toolbelt==0.9.1
requests==2.22.0
responder==2.0.4
rfc3986==1.3.2
rx==1.6.1
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
smart-open==1.9.0
starlette==0.12.13
typed-ast==1.4.0
typesystem==0.2.4
typing-extensions==3.7.4.1
urllib3==1.25.7 ; python_version >= '3.4'
uvicorn==0.8.6
uvloop==0.12.2 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'pypy'
websockets==7.0
whitenoise==4.1.4
